\name{CategoricalDataAnalysis-package}
\alias{CategoricalDataAnalysis-package}
\alias{CategoricalDataAnalysis}
\docType{package}
\title{
\packageTitle{CategoricalDataAnalysis}
}
\description{
\packageDescription{CategoricalDataAnalysis}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{CategoricalDataAnalysis}
\packageIndices{CategoricalDataAnalysis}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{CategoricalDataAnalysis}

Maintaitexlive-fonts-extraner: \packageMaintainer{CategoricalDataAnalysis}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation directory ~~
\keyword{ package }
\seealso{

}
\examples{

data("crabs2")
catbarchart(continous2categorical(crabs2))

}
