\name{CategoricalDataAnalysis-package}
\alias{CategoricalDataAnalysis-package}
\alias{CategoricalDataAnalysis}
\docType{package}
\title{
\packageTitle{CategoricalDataAnalysis}
}
\description{
\packageDescription{CategoricalDataAnalysis}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{CategoricalDataAnalysis}
\packageIndices{CategoricalDataAnalysis}

}
\author{
\packageAuthor{CategoricalDataAnalysis}

Maintaitexlive-fonts-extraner: \packageMaintainer{CategoricalDataAnalysis}
}
\references{
}
\keyword{ package }
\seealso{

}
\examples{

data("crabs2")
catbarchart(continous2categorical(crabs2))

}
